Pimped the String editor. Added OnFocus.
Each time a editor is brought to the top of the editor stack it's OnFocus function is called.
This is usefull for when you have a task that has to be done each time this happens.
An example is when using the string editor, after storing the resulting string and you want to use the result.