Added caching to SavableManager.
As said, a 1 minute fix.
Note: this is -not- an attempt at optimazation.
This is required because several places may ask a handle to the same database row.
For example, if two Rooms request a handle to a character, it is expected that when they change a field on the character this change is visible, regardless of which handle is used to access the object.